requirements.txt
lyzr
streamlit
openai
python-dotenv